Type check added, and basics exception added for all CRUD operations.

+++ src/main/java/com/pavelDinit/dinitProject/services/UrlService.java
@@ -2,6 +2,8 @@ package com.pavelDinit.dinitProject.services;
 
 import com.pavelDinit.dinitProject.dtos.UrlCreationDto;
 import com.pavelDinit.dinitProject.dtos.UrlReadingDto;
+import com.pavelDinit.dinitProject.exceptions.badrequest.ApiBadRequest;
+import com.pavelDinit.dinitProject.exceptions.badrequest.TypeMissmatch;
 import com.pavelDinit.dinitProject.exceptions.conflict.Conflict;
 import com.pavelDinit.dinitProject.exceptions.notfound.ResourceNotFound;
 import com.pavelDinit.dinitProject.models.Url;
@@ -25,7 +27,6 @@ public class UrlService {
     Logger logger = LoggerFactory.getLogger(UrlService.class);
 
 
-
     public UrlService(UrlRepo urlRepo, RestTemplate restTemplate) {
         this.urlRepo = urlRepo;
         this.restTemplate = restTemplate;
@@ -43,24 +44,28 @@ public class UrlService {
 
 
     // Function to get a single URL by its name:
-    public UrlReadingDto getUrlById(Long urlId){
-        Url url = urlRepo.findByUrlId(urlId)
-                .orElseThrow(() -> new ResourceNotFound("URL with name " + urlId + " not found."));
+    public UrlReadingDto getUrlById(Long urlId) {
+
+        if (urlId == null || urlId <= 0) {
+            throw new TypeMissmatch("Invalid URL ID: " + urlId);
+        }
+
+        Url url = urlRepo.findByUrlId(urlId).orElseThrow(() -> new ResourceNotFound("URL with name " + urlId + " not found."));
         return UrlReadingDto.readingDtoFromUrl(url);
     }
 
     // Function to get all the names of the URLS that are inside the URL list:
-    public List<Object> getAllUrlNames(){
+    public List<Object> getAllUrlNames() {
         List<UrlReadingDto> all = getAllUrls();
-        if(all.isEmpty()){
+        if (all.isEmpty()) {
             throw new ResourceNotFound("There are no urls stored at the moment.");
         }
-            return all.stream().map(UrlReadingDto::getUrlName).collect(Collectors.toList());
+        return all.stream().map(UrlReadingDto::getUrlName).collect(Collectors.toList());
     }
 
     // Function to delete a single URL based on its ID:
     public String deleteUrlById(Long urlId) {
-        if(!urlRepo.existsById(urlId)){
+        if (!urlRepo.existsById(urlId)) {
             throw new ResourceNotFound("There is no url with id " + urlId + ".");
         }
         urlRepo.deleteByUrlId(urlId);
@@ -69,33 +74,37 @@ public class UrlService {
 
     // Function to delete all URLs:
     public String deleteAllUrls() {
+        List<Url> allUrls = urlRepo.findAll();
+        if (allUrls.isEmpty()) {
+            throw new ResourceNotFound("There are no URLs stored at the moment.");
+        }
         urlRepo.deleteAll();
         return "All URLs have been deleted.";
     }
+
 
     // Function to add a single URL:
     // This one will probably read some info that users would give
     // And then write that one:
     public String addUrl(UrlCreationDto urlCreateDTO) {
 
-        // Check full urls:
+    // Check full urls:
         if (urlCreateDTO.getFullUrl() == null || urlCreateDTO.getFullUrl().isEmpty()) {
             throw new Conflict("There is no URL specified.");
         } else if (!UrlReadingDto.checkUrlValidity(urlCreateDTO.getFullUrl())) {
             throw new Conflict("Invalid URL format.");
-        }
-
-        // Check name:
-        if (urlCreateDTO.getUrlName() == null || urlCreateDTO.getUrlName().isEmpty()) {
+        } else if (urlCreateDTO.getUrlName() == null || urlCreateDTO.getUrlName().isEmpty()) {
             throw new Conflict("There is no URL name specified.");
-        }
+        } else if (urlCreateDTO.getAddedByUserId() == null) {
+            throw new Conflict("There is no User Id specified.");
+            // This will later be used for authentication.
+    }
 
         // Check if url exists:
         Optional<Url> existingUrl = urlRepo.findByFullUrl(urlCreateDTO.getFullUrl());
         if (existingUrl.isPresent()) {
             throw new Conflict("URL already exists.");
         }
-
 
         String urlHealthStr = urlCreateDTO.getFullUrl();
         boolean urlHealth = checkUrlHealth1(urlHealthStr);
@@ -105,8 +114,8 @@ public class UrlService {
         return "Created URL with ID: " + url.getUrlId();
     }
 
-    // Function for checking if an url is healthy or not.
-    // Checks inside for Healthy status otherwise it will give us false.
+// Function for checking if an url is healthy or not.
+// Checks inside for Healthy status otherwise it will give us false.
 //    public boolean checkUrlHealth(String fullUrl) {
 //        try {
 //            ResponseEntity<String> response = restTemplate.getForEntity(fullUrl, String.class);
@@ -123,7 +132,7 @@ public class UrlService {
 //    }
 
     // Function that checks if we are receiving 200 from an url, if we are it will give 200
-    // This is for testing purposes, later we switch to checkUrlHealth function that is currently commented above.
+// This is for testing purposes, later we switch to checkUrlHealth function that is currently commented above.
     public boolean checkUrlHealth1(String fullUrl) {
         try {
             ResponseEntity<String> response = restTemplate.getForEntity(fullUrl, String.class);
@@ -135,9 +144,16 @@ public class UrlService {
         }
     }
 
+    // Function to return if url is healthy or not:
+    public boolean checkUrlHealthById(Long id) {
+        UrlReadingDto url = getUrlById(id);
+        String fullUrl = url.getFullUrl();
+        return checkUrlHealth1(fullUrl);
+    }
+
 
     // Function to check and update all the URLS healths, im not sure how to write this XD
-    public void checkAllUrlsHealth(){
+    public void checkAllUrlsHealth() {
         List<Url> urls = urlRepo.findAll();
 
         if (urls.isEmpty()) {
